requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.12.3
bleach==6.2.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.9
decorator==5.1.1
Django==5.1.6
django-cors-headers==4.7.0
django-markdownx==4.0.9
django-summernote==0.8.20.0
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.1.0
frozenlist==1.8.0
genai==2.1.0
google-api-core==2.25.2
google-auth==2.41.1
google-genai==1.41.0
googleapis-common-protos==1.70.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
load-dotenv==0.1.0
lxml==6.0.0
Markdown==3.8.2
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.2.2
openai==0.27.10
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
postgres==4.0
prompt_toolkit==3.0.48
propcache==0.4.0
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyairtable==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.1
pyzmq==26.2.0
regex==2025.9.18
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.3.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xlsxwriter==3.2.5
yarl==1.21.0
boto3>=1.34.0
django-ses>=3.0.0
PyJWT>=2.8.0
djangorestframework==3.16.1
